move intercept logic
