+++ .nycrc.json
@@ -1,6 +1,4 @@
 {
-  "exclude": [
-    "support-utils.js",
-    "task-utils.js"
-  ]
+  "exclude": ["support-utils.js", "task-utils.js"],
+  "reporter": ["html", "text"]
 }
