+++ cypress/e2e/calculator.cy.js
@@ -1,53 +1,9 @@
 import { CalculatorPage } from './calculator-po'
 
-// @ts-ignore
-import { createInstrumenter } from 'istanbul-lib-instrument'
-
-const instrumenter = createInstrumenter({
-  esModules: true,
-  compact: false,
-  preserveComments: true,
-})
-
-const baseUrl = Cypress.config('baseUrl')
-// @ts-ignore
-const proxyServer = Cypress.config('proxyServer') + '/'
-
-const instrumentScripts = Cypress.env('coverage')?.instrument
-
 it(
   'collects code coverage on the fly',
   { viewportWidth: 1200, viewportHeight: 1000, baseUrl: null },
   () => {
-    if (instrumentScripts) {
-      cy.intercept(
-        {
-          method: 'GET',
-          resourceType: 'script',
-          url: instrumentScripts,
-        },
-        (req) => {
-          // TODO: remove caching headers for now
-          // @ts-ignore
-          req.continue((res) => {
-            const relativeUrl = req.url
-              // @ts-ignore
-              .replace(baseUrl, '')
-              .replace(proxyServer, '')
-            // console.log('instrumenting', relativeUrl)
-
-            // @ts-ignore
-            const instrumented = instrumenter.instrumentSync(
-              res.body,
-              relativeUrl,
-            )
-            res.body = instrumented
-            return res
-          })
-        },
-      )
-    }
-
     cy.visit('cypress/calculator/index.html')
     // after instrumenting the coverage should be collected in the window object
     // under window.__coverage__ key
